feat: rendering cylindrical constraints
Currently, the cylinders are mesh_drawables, and we between each frame clear the data in the
mesh_drawable which is on the GPU, then we re-initialise it. Might be a better way of doing this
for cylinders the position/rotation transformations for this struct is quite strange in terms
of the expected result. Further, comments have been added to indicate which joint is which
when we define sphere + cylindrical constraints.